requirements.txt
Django==5.2.7
PyMuPDF==1.26.0
Pillow==11.2.1
django-allauth==65.8.1
